requirements.txt
Django >= 1.4
djangorestframework>=2.3,<3
hashids
